# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfólio API Docs",
        "description": "API para consulta de dados do portfólio",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença API",
            "url": "https://github.com/felipesilva15/portfolio-manager-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/technology": {
            "get": {
                "tags": [
                    "Technology"
                ],
                "summary": "List all technologys",
                "operationId": "f8835b096f343ccbb1f912bd7783cf3a",
                "responses": {
                    "200": {
                        "description": "Technology list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Technology"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Technology"
                ],
                "summary": "Registers a technology",
                "operationId": "88076b9969bc2f97fa9b7a03e7a2d09b",
                "requestBody": {
                    "description": "Data for creating a new technology",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TechnologyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered technology data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Technology"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/technology/{id}": {
            "get": {
                "tags": [
                    "Technology"
                ],
                "summary": "List a technology by ID",
                "operationId": "9818ae0e65e880d40c7db37fa9843d2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Technology ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Technology data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Technology"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Technology"
                ],
                "summary": "Update a technology",
                "operationId": "604287e92fe1a99eb0911dd965698f6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Technology ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update technology",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TechnologyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated technology data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Technology"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Technology"
                ],
                "summary": "Delete a technology",
                "operationId": "4450a703ea579556bf26436d49adba68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Technology ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorDTO": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "/api/route"
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "Error ocurried"
                    }
                },
                "type": "object"
            },
...
            "TechnologyRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 80,
                        "minLength": 1,
                        "example": "Laravel"
                    }
                },
                "type": "object"
            },
...
            "Technology": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laravel"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}